fixed the infinite request

+++ lib/features/pos/layouts/restaurant_layout.dart
@@ -1082,7 +1082,7 @@ class _RestaurantLayoutState extends BasePosLayoutState<RestaurantLayout> {
               ),
             ),
             onPressed: () {
-              final qty = int.tryParse(controller.text) ?? 1;
+              final qty = double.tryParse(controller.text) ?? 1.0;
               widget.cartNotifier.updateQuantity(item.product, qty);
               Navigator.pop(ctx);
             },
